+++ processor/src/main/java/router/fu/processor/ProcessorUtil.java
@@ -1,51 +1,18 @@
 package router.fu.processor;
 
-import com.squareup.javapoet.AnnotationSpec;
-import com.squareup.javapoet.MethodSpec;
-import com.squareup.javapoet.ParameterSpec;
-import java.lang.annotation.Documented;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
-import javax.lang.model.AnnotatedConstruct;
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.AnnotationValue;
 import javax.lang.model.element.Element;
-import javax.lang.model.type.DeclaredType;
 import org.realityforge.proton.AnnotationsUtil;
 
 final class ProcessorUtil
 {
   private ProcessorUtil()
   {
-  }
-
-
-  static void copyDocumentedAnnotations( @Nonnull final AnnotatedConstruct element,
-                                         @Nonnull final MethodSpec.Builder builder )
-  {
-    for ( final AnnotationMirror annotation : element.getAnnotationMirrors() )
-    {
-      final DeclaredType annotationType = annotation.getAnnotationType();
-      if ( !annotationType.toString().startsWith( "router.fu.annotations." ) &&
-           null != annotationType.asElement().getAnnotation( Documented.class ) )
-      {
-        builder.addAnnotation( AnnotationSpec.get( annotation ) );
-      }
-    }
-  }
-
-  static void copyDocumentedAnnotations( @Nonnull final AnnotatedConstruct element,
-                                         @Nonnull final ParameterSpec.Builder builder )
-  {
-    for ( final AnnotationMirror annotation : element.getAnnotationMirrors() )
-    {
-      if ( null != annotation.getAnnotationType().asElement().getAnnotation( Documented.class ) )
-      {
-        builder.addAnnotation( AnnotationSpec.get( annotation ) );
-      }
-    }
   }
 
   @SuppressWarnings( "unchecked" )