+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Use the `org.realityforge.proton:proton-processor-pack` project for processor utility methods.
 * Shade the `auto-common` artifact in the annotation processor.
 * Use builtin `SourceVersion.isIdentifier(String)` method rather than custom code that emulates behaviour.
+* When generating the router, stop copying all `@Documented` annotated annotations and instead copy the specific set of whitelisted annotations that developers typically care about. These include the nullability annotations `@Nonnull` and `@Nullable` as well ass `@Deprecated`. 
 
 ### [v0.21](https://github.com/realityforge/router-fu/tree/v0.21) (2019-12-17) · [Full Changelog](https://github.com/realityforge/router-fu/compare/v0.20...v0.21)